// libs/modrinth/src/lib/types.ts
export type ModrinthId = string;

export type Environment = "required" | "optional" | "unsupported" | "unknown";
// ...
export type ProjectType = "mod" | "modpack" | "resourcepack" | "shader";
export type MonetizationStatus = "monetized" | "demonetized" | "force-demonetized";

export type GameVersion = string;
export type Platform = string;
export type Category = string;
export type CategoryOrPlatform = Category | Platform;
// ...
export interface SearchResult {
  id: ModrinthId;
  project_id: ModrinthId;
  project_type: ProjectType;
  slug: string;
  title: string;
  description: string;
  monetization_status: MonetizationStatus;

  icon_url?: string;
  color?: number;

  categories: CategoryOrPlatform[];
  display_categories: CategoryOrPlatform[];
  versions: GameVersion[];
  latest_version: GameVersion;

  downloads: number;
  follows: number;

  client_side: Environment;
  server_side: Environment;

  author: string;

  date_created: string;
  date_modified: string;

  gallery: string[];
  featured_gallery?: string[];

  license: string;
}